requirements.txt
notion-client>=2.0.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0 